Report Tempo tool changes as the tempo the ratio encodes
`TempoChange::getAbsoluteTempo` returned `int(std::lround(...))`, costing
up to half a beat per minute. A JW Tempo ramp in denigma's
`tempo_changes.musx` converts as 50, 53, 57, 60 where the file encodes
50, 53.312, 56.624, 59.937.

Return `double` instead. `ratio` is an integer, so it cannot encode most
tempos exactly: a quarter-note tempo of 50 at unit 1000 is stored as
55924, whose exact quotient is 49.99995. Every tempo that converts back
to the same ratio states the instance equally well, so report the
shortest of them. A stored 50 then reads as 50 while a genuinely
fractional step keeps its own value, and no precision is invented, since
a decimal too fine for the ratio to distinguish never converts back.

This is not shortest-round-trip double printing. The grid is the integer
ratio, about 0.0009 bpm at a quarter note, not the mantissa, so a dtoa
routine cannot do this job.

The existing test's whole-note case moves from 4 to 3.75, its exact
quotient being 3.74995. Adds a case over the ratios from the ramp above,
and a property test running the conversion and its inverse over ~170
ratios at 5 note types.

+++ src/musx/dom/Others.h
@@ -2993,9 +2993,12 @@ public:
 
     /// @brief Computes the absolute tempo represented by the TempoChange Instance.
     /// @param noteType optional note type for which to get the beats per minute. (The default is NoteType::Quarter.)
-    /// @return The per minute value.
+    /// @return The per minute value. Because #ratio is an integer, it cannot encode most tempos exactly.
+    /// The value returned is the shortest decimal that converts back to #ratio, which states the encoded
+    /// tempo without adding precision the encoding does not carry. It therefore usually has a fractional
+    /// part, and callers that need a whole number must round it themselves.
     /// @throws std::logic_error if this instance represents a relative tempo.
-    int getAbsoluteTempo(NoteType noteType = NoteType::Quarter) const;
+    double getAbsoluteTempo(NoteType noteType = NoteType::Quarter) const;
 
     constexpr static std::string_view XmlNodeName = "tempoDef"; ///< The XML node name for this type.
     static const xml::XmlElementArray<TempoChange>& xmlMappingArray(); ///< Required for musx::factory::FieldPopulator.
